// src/imports/api.ts
export interface ApiPlayerIdGoalsGet200Response {
}
export interface ApiPlayerIdRatingsGet200ResponseInner {
}
// ...
export interface CreateMatchDTOAttendance {
}
// ...
export interface GetCompetitionDTO {
    'id': string;
    'name': string;
    'country': string;
    'seasons': Array<GetSeasonDTO> | null;
    'matches': Array<GetMatchDTO> | null;
}
export interface GetMatchDTO {
    'id': string;
    'kickOff': string;
    'venue': string;
    'referee': string;
    'attendance': CreateMatchDTOAttendance | null;
    'homeGoals': ApiPlayerIdGoalsGet200Response;
    'awayGoals': ApiPlayerIdGoalsGet200Response;
    'status': string;
    'homeTeam': GetTeamDTO;
    'awayTeam': GetTeamDTO;
    'competition': GetCompetitionDTO;
    'season': GetSeasonDTO;
}
export interface GetPlayerDTO {
    'id': string;
    'firstName': string;
    'lastName': string;
    'position': string;
    'alternativePositions': Array<string> | null;
    'shirtNumber': ApiPlayerIdGoalsGet200Response;
    'dateOfBirth': string;
    'national': string;
    'isCaptain': boolean;
    'isActive': boolean;
    'team': GetTeamDTO | null;
}
export interface GetPlayerMatchStatsDTO {
    'id': string;
    'match': GetMatchDTO;
    'player': GetPlayerDTO;
    'team': GetTeamDTO;
    'started': boolean;
    'wasSubstitutedOn': boolean;
    'wasSubstitutedOff': boolean;
    'isCaptain': boolean;
    'isManOfTheMatch': boolean;
    'analysis': PlayerAnalysis | null;
    'playerStats': PlayerStats;
    'playerAttack': PlayerAttack;
    'playerPasses': PlayerPasses;
    'playerDefence': PlayerDefence;
    'playerDuels': PlayerDuels;
    'goalkeepering': Goalkeepering;
    'playerDiscipline': PlayerDiscipline;
}
export interface GetSeasonDTO {
    'id': string;
    'name': string;
    'startDate': string;
    'endDate': string;
    'isCurrent': boolean;
}
// ...
export interface GetTeamDTO {
    'id': string;
    'name': string;
    'shortName': string;
    'stadium': string;
    'city': string;
    'foundedYear': ApiPlayerIdGoalsGet200Response;
    'coach': string;
    'captain': GetPlayerDTO | null;
    'preferredFormation': string | null;
    'playingStyle': string | null;
}
export interface GetTeamMatchStatsDTO {
    'id': string;
    'match': GetMatchDTO;
    'team': GetTeamDTO;
    'isHome': boolean;
    'formation': string;
    'playingStyle': string;
    'matchStats': MatchStats;
    'matchShots': MatchShots;
    'matchExpectedGoals': MatchExpectedGoals;
    'matchPasses': MatchPasses;
    'matchDiscipline': MatchDiscipline;
    'matchDefence': MatchDefence;
    'matchDuels': MatchDuels;
    'matchAttackingZones': MatchAttackingZones;
    'matchAnalysis': MatchAnalysis | null;
}
// ...
export interface Goalkeepering {
    'saves'?: ApiPlayerIdGoalsGet200Response;
    'goalsConceded'?: ApiPlayerIdGoalsGet200Response;
    'facedxGOT'?: ApiPlayerIdRatingsGet200ResponseInner;
    'goalsPrevented'?: ApiPlayerIdRatingsGet200ResponseInner;
    'actedAsSweeper'?: ApiPlayerIdGoalsGet200Response;
    'highClaim'?: ApiPlayerIdGoalsGet200Response;
    'longBalls'?: ApiPlayerIdGoalsGet200Response;
    'accurateLongBalls'?: ApiPlayerIdGoalsGet200Response;
    'passes'?: ApiPlayerIdGoalsGet200Response;
    'accuratePasses'?: ApiPlayerIdGoalsGet200Response;
}
export interface GoalsPerRoundDTO {
    'date': string;
    'goals': ApiPlayerIdGoalsGet200Response;
}
export interface MatchAnalysis {
    'tacticalNotes'?: string;
    'analystNotes'?: string;
}
export interface MatchAttackingZones {
    'centerAttack'?: ApiPlayerIdGoalsGet200Response;
    'leftAttack'?: ApiPlayerIdGoalsGet200Response;
    'rightAttack'?: ApiPlayerIdGoalsGet200Response;
}
export interface MatchDefence {
    'tackles'?: ApiPlayerIdGoalsGet200Response;
    'interceptions'?: ApiPlayerIdGoalsGet200Response;
    'blocks'?: ApiPlayerIdGoalsGet200Response;
    'clearances'?: ApiPlayerIdGoalsGet200Response;
    'keeperSaves'?: ApiPlayerIdGoalsGet200Response;
}
export interface MatchDiscipline {
    'foulsCommitted'?: ApiPlayerIdGoalsGet200Response;
    'yellowCards'?: ApiPlayerIdGoalsGet200Response;
    'redCards'?: ApiPlayerIdGoalsGet200Response;
}
export interface MatchDuels {
    'duelsWon'?: ApiPlayerIdGoalsGet200Response;
    'groundDuelsWon'?: ApiPlayerIdGoalsGet200Response;
    'groundDuelsWonPercentage'?: ApiPlayerIdGoalsGet200Response;
    'aerialDuelsWon'?: ApiPlayerIdGoalsGet200Response;
    'aerialDuelsWonPercentage'?: ApiPlayerIdGoalsGet200Response;
    'successfulDribbles'?: ApiPlayerIdGoalsGet200Response;
    'successfulDribblesPercentage'?: ApiPlayerIdGoalsGet200Response;
}
export interface MatchExpectedGoals {
    'xg'?: ApiPlayerIdRatingsGet200ResponseInner;
    'openPlayXG'?: ApiPlayerIdRatingsGet200ResponseInner;
    'setPlayXG'?: ApiPlayerIdRatingsGet200ResponseInner;
    'nonPenaltyXG'?: ApiPlayerIdRatingsGet200ResponseInner;
    'xgot'?: ApiPlayerIdRatingsGet200ResponseInner;
}
export interface MatchPasses {
    'passes'?: ApiPlayerIdGoalsGet200Response;
    'accuratePasses'?: ApiPlayerIdGoalsGet200Response;
    'ownHalf'?: ApiPlayerIdGoalsGet200Response;
    'oppositionHalf'?: ApiPlayerIdGoalsGet200Response;
    'accurateLongBalls'?: ApiPlayerIdGoalsGet200Response;
    'accurateLongBallsPercentage'?: ApiPlayerIdGoalsGet200Response;
    'accurateCrosses'?: ApiPlayerIdGoalsGet200Response;
    'accurateCrossesPercentage'?: ApiPlayerIdGoalsGet200Response;
    'throws'?: ApiPlayerIdGoalsGet200Response;
    'touchesInOppositionBox'?: ApiPlayerIdGoalsGet200Response;
    'offsides'?: ApiPlayerIdGoalsGet200Response;
}
export interface MatchShots {
    'totalShots'?: ApiPlayerIdGoalsGet200Response;
    'shotsOnTarget'?: ApiPlayerIdGoalsGet200Response;
    'shotsOffTarget'?: ApiPlayerIdGoalsGet200Response;
    'blockedShots'?: ApiPlayerIdGoalsGet200Response;
    'hitWoodwork'?: ApiPlayerIdGoalsGet200Response;
    'shotsInsideBox'?: ApiPlayerIdGoalsGet200Response;
    'shotsOutsideBox'?: ApiPlayerIdGoalsGet200Response;
}
export interface MatchStats {
    'teamGoals'?: ApiPlayerIdGoalsGet200Response;
    'oppositionGoals'?: ApiPlayerIdGoalsGet200Response;
    'possession'?: ApiPlayerIdGoalsGet200Response;
    'corners'?: ApiPlayerIdGoalsGet200Response;
    'bigChances'?: ApiPlayerIdGoalsGet200Response;
    'bigChancesMissed'?: ApiPlayerIdGoalsGet200Response;
}
export interface MatchSummaryDTO {
    'match': GetMatchDTO;
    'scoreline': string;
    'homeTeamStats': GetTeamMatchStatsDTO | null;
    'awayTeamStats': GetTeamMatchStatsDTO | null;
    'topRatedPlayers': Array<GetPlayerMatchStatsDTO>;
    'goalScorers': Array<GetPlayerMatchStatsDTO>;
    'assistProviders': Array<GetPlayerMatchStatsDTO>;
}
export interface MatchWorkspaceDTO {
    'match': GetMatchDTO;
    'homeTeam': GetTeamDTO;
    'awayTeam': GetTeamDTO;
    'competition': GetCompetitionDTO;
    'season': GetSeasonDTO;
    'homeTeamStats': GetTeamMatchStatsDTO | null;
    'awayTeamStats': GetTeamMatchStatsDTO | null;
    'homePlayers': Array<GetPlayerMatchStatsDTO>;
    'awayPlayers': Array<GetPlayerMatchStatsDTO>;
}
export interface PlayerAnalysis {
    'performanceSummary'?: string;
    'analystNotes'?: string;
}
export interface PlayerAttack {
    'goals'?: ApiPlayerIdGoalsGet200Response;
    'xg'?: ApiPlayerIdRatingsGet200ResponseInner;
    'xgot'?: ApiPlayerIdRatingsGet200ResponseInner;
    'totalShots'?: ApiPlayerIdGoalsGet200Response;
    'shotsOnTarget'?: ApiPlayerIdGoalsGet200Response;
    'touchesInOppositionBox'?: ApiPlayerIdGoalsGet200Response;
    'bigChancesMissed'?: ApiPlayerIdGoalsGet200Response;
    'successfulDribbles'?: ApiPlayerIdGoalsGet200Response;
    'dribblesAttempted'?: ApiPlayerIdGoalsGet200Response;
}
export interface PlayerComparisonDTO {
    'player': PlayerStatisticsDTO;
    'otherPlayer': PlayerStatisticsDTO;
}
export interface PlayerDefence {
    'defensiveContributions'?: ApiPlayerIdGoalsGet200Response;
    'tackles'?: ApiPlayerIdGoalsGet200Response;
    'interceptions'?: ApiPlayerIdGoalsGet200Response;
    'blocks'?: ApiPlayerIdGoalsGet200Response;
    'recoveries'?: ApiPlayerIdGoalsGet200Response;
    'clearance'?: ApiPlayerIdGoalsGet200Response;
    'headedClearances'?: ApiPlayerIdGoalsGet200Response;
    'dribbledPast'?: ApiPlayerIdGoalsGet200Response;
}
export interface PlayerDiscipline {
    'yellowCards'?: ApiPlayerIdGoalsGet200Response;
    'redCards'?: ApiPlayerIdGoalsGet200Response;
    'foulsCommitted'?: ApiPlayerIdGoalsGet200Response;
    'wasFouled'?: ApiPlayerIdGoalsGet200Response;
}
export interface PlayerDuels {
    'duelsWon'?: ApiPlayerIdGoalsGet200Response;
    'duelsLost'?: ApiPlayerIdGoalsGet200Response;
    'groundDuelsWon'?: ApiPlayerIdGoalsGet200Response;
    'totalGroundDuels'?: ApiPlayerIdGoalsGet200Response;
    'aerialDuelsWon'?: ApiPlayerIdGoalsGet200Response;
    'totalAerialDuels'?: ApiPlayerIdGoalsGet200Response;
}
export interface PlayerLeaderDTO {
    'playerId': string;
    'firstName': string;
    'lastName': string;
    'teamName': string;
    'total': ApiPlayerIdGoalsGet200Response;
    'averageRating': ApiPlayerIdRatingsGet200ResponseInner;
}
export interface PlayerPasses {
    'touches'?: ApiPlayerIdGoalsGet200Response;
    'accuratePasses'?: ApiPlayerIdGoalsGet200Response;
    'passesAttempted'?: ApiPlayerIdGoalsGet200Response;
    'assists'?: ApiPlayerIdGoalsGet200Response;
    'xa'?: ApiPlayerIdRatingsGet200ResponseInner;
    'chancesCreated'?: ApiPlayerIdGoalsGet200Response;
    'passesIntoFinalThird'?: ApiPlayerIdGoalsGet200Response;
    'accurateCrosses'?: ApiPlayerIdGoalsGet200Response;
    'crossesAttempted'?: ApiPlayerIdGoalsGet200Response;
    'accurateLongBalls'?: ApiPlayerIdGoalsGet200Response;
    'longBallsAttempted'?: ApiPlayerIdGoalsGet200Response;
}
// ...
export interface PlayerStats {
    'fotmobRating'?: ApiPlayerIdRatingsGet200ResponseInner;
    'sofascoreRating'?: ApiPlayerIdRatingsGet200ResponseInner;
    'minutesPlayed'?: ApiPlayerIdGoalsGet200Response;
}